fix: restructuring audio-call game files
