+++ src/app/app.ts
@@ -2,7 +2,8 @@ import Navigo from "navigo";
 import { RENDER_BASIC_STRUCTURE } from "../pages/main/main-page-render";
 import { COMPONENT_HEADER } from "../components/header/header";
 import { COMPONENT_FOOTER } from "../components/footer/footer";
-import { AUDIO_CALL_GAME } from "../pages/audio-call-game/audio-game";
+import { AUDIO_CALL_RENDER } from "../pages/audio-call-game/audio-game-render";
+import { AUDIO_CALL_GAME } from "../controller/audio-call-game/audio-call-game";
 
 const router = new Navigo("/", { hash: true });
 
@@ -16,7 +17,7 @@ router.on("/", async () => {
 router.on("/audio-call-game", async () => {
   COMPONENT_HEADER.createHeader();
   COMPONENT_FOOTER.createFooter();
-  AUDIO_CALL_GAME.renderRules();
+  AUDIO_CALL_RENDER.renderRules();
   AUDIO_CALL_GAME.AudioCallListenerStart();
 });
